+++ tartiflette/schema/bakery.py
@@ -9,14 +9,16 @@ from tartiflette.schema import GraphQLSchema
 
 class SchemaBakery:
     @staticmethod
-    def _preheat(schema_name):
+    def _preheat(schema_name, exclude_builtins_scalars):
         schema_info = SchemaRegistry.find_schema_info(schema_name)
         schema = schema_info.get("inst", GraphQLSchema(name=schema_name))
 
         sdl = schema_info["sdl"]
         build_graphql_schema_from_sdl(sdl, schema=schema)
 
-        SchemaBakery._inject_default_object(schema_name)
+        SchemaBakery._inject_default_object(
+            schema_name, exclude_builtins_scalars
+        )
 
         for object_ids in ["directives", "resolvers", "scalars"]:
             for obj in schema_info.get(object_ids, []):
@@ -27,17 +29,24 @@ class SchemaBakery:
         return schema
 
     @staticmethod
-    def bake(schema_name, custom_default_resolver=None):
-        schema = SchemaBakery._preheat(schema_name)
+    def bake(
+        schema_name,
+        custom_default_resolver=None,
+        exclude_builtins_scalars=None,
+    ):
+        schema = SchemaBakery._preheat(schema_name, exclude_builtins_scalars)
         schema.bake(custom_default_resolver)
-
         return schema
 
     @staticmethod
-    def _inject_default_object(schema_name):
+    def _inject_default_object(schema_name, exclude_builtins_scalars):
         for name, scalar_implem in CUSTOM_SCALARS.items():
-            deco = Scalar(name, schema_name)
-            deco(scalar_implem)
+            if (
+                exclude_builtins_scalars is None
+                or name not in exclude_builtins_scalars
+            ):
+                deco = Scalar(name, schema_name)
+                deco(scalar_implem)
 
         for name, directive_implem in BUILT_IN_DIRECTIVES.items():
             deco = Directive(name, schema_name)