+++ tests/functional/test_introspection.py
@@ -233,14 +233,121 @@ async def test_tartiflette_execute_schema_introspection_output(clean_registry):
                     {"kind": "OBJECT", "name": "CustomRootQuery"},
                     {"kind": "OBJECT", "name": "CustomRootMutation"},
                     {"kind": "OBJECT", "name": "CustomRootSubscription"},
-                    {"kind": "SCALAR", "name": "Int"},
-                    {"kind": "SCALAR", "name": "String"},
-                    {"kind": "SCALAR", "name": "ID"},
-                    {"kind": "SCALAR", "name": "Float"},
                     {"kind": "SCALAR", "name": "Boolean"},
                     {"kind": "SCALAR", "name": "Date"},
-                    {"kind": "SCALAR", "name": "Time"},
                     {"kind": "SCALAR", "name": "DateTime"},
+                    {"kind": "SCALAR", "name": "Float"},
+                    {"kind": "SCALAR", "name": "ID"},
+                    {"kind": "SCALAR", "name": "Int"},
+                    {"kind": "SCALAR", "name": "String"},
+                    {"kind": "SCALAR", "name": "Time"},
+                    {"kind": "OBJECT", "name": "__Schema"},
+                    {"kind": "ENUM", "name": "__TypeKind"},
+                    {"kind": "OBJECT", "name": "__Type"},
+                    {"kind": "OBJECT", "name": "__Field"},
+                    {"kind": "OBJECT", "name": "__InputValue"},
+                    {"kind": "OBJECT", "name": "__EnumValue"},
+                    {"name": "__DirectiveLocation", "kind": "ENUM"},
+                    {"kind": "OBJECT", "name": "__Directive"},
+                ],
+                "directives": [
+                    {
+                        "description": None,
+                        "locations": ["FIELD_DEFINITION", "ENUM_VALUE"],
+                        "args": [
+                            {
+                                "name": "reason",
+                                "type": {"kind": "SCALAR", "name": "String"},
+                                "description": None,
+                                "defaultValue": "No longer supported",
+                            }
+                        ],
+                        "name": "deprecated",
+                    },
+                    {
+                        "args": [],
+                        "name": "non_introspectable",
+                        "description": None,
+                        "locations": ["FIELD_DEFINITION"],
+                    },
+                ],
+                "queryType": {"name": "CustomRootQuery"},
+                "mutationType": {"name": "CustomRootMutation"},
+                "subscriptionType": {"name": "CustomRootSubscription"},
+            }
+        }
+    } == result
+
+
+@pytest.mark.asyncio
+async def test_tartiflette_execute_schema_introspection_output_exclude_scalars(
+    clean_registry
+):
+    schema_sdl = """
+    schema {
+        query: CustomRootQuery
+        mutation: CustomRootMutation
+        subscription: CustomRootSubscription
+    }
+
+    type CustomRootQuery {
+        test: String
+    }
+
+    type CustomRootMutation {
+        test: Int
+    }
+
+    type CustomRootSubscription {
+        test: String
+    }
+    """
+
+    ttftt = Engine(schema_sdl, exclude_builtins_scalars=["Date", "DateTime"])
+
+    result = await ttftt.execute(
+        """
+    query Test{
+        __schema {
+            queryType { name }
+            mutationType { name }
+            subscriptionType { name }
+            types {
+                kind
+                name
+            }
+            directives {
+                name
+                description
+                locations
+                args {
+                    name
+                    description
+                    type {
+                        kind
+                        name
+                    }
+                    defaultValue
+                }
+            }
+        }
+    }
+    """
+    )
+
+    assert {
+        "data": {
+            "__schema": {
+                "types": [
+                    {"kind": "OBJECT", "name": "CustomRootQuery"},
+                    {"kind": "OBJECT", "name": "CustomRootMutation"},
+                    {"kind": "OBJECT", "name": "CustomRootSubscription"},
+                    {"kind": "SCALAR", "name": "Boolean"},
+                    {"kind": "SCALAR", "name": "Float"},
+                    {"kind": "SCALAR", "name": "ID"},
+                    {"kind": "SCALAR", "name": "Int"},
+                    {"kind": "SCALAR", "name": "String"},
+                    {"kind": "SCALAR", "name": "Time"},
                     {"kind": "OBJECT", "name": "__Schema"},
                     {"kind": "ENUM", "name": "__TypeKind"},
                     {"kind": "OBJECT", "name": "__Type"},