+++ CHANGELOG.md
@@ -5,13 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Release]
+
+### [0.2.2] - 2019-01-04
 
 ### Added
 
+- Enable you to exclude builtins scalars at Engine initialization time with `exclude_builtins_scalars` parameter.
 - Publish package to Test PyPi on each working branch.
 
-## [Release]
+```python
+e = Engine("my_sdl.sdl", exclude_builtins_scalars=["Date", "DateTime"])
+```
 
 ### [0.2.1] - 2018-12-27